divided business logic and ui in chart widgets
