+++ packages/custom_fl_chart/lib/custom_fl_chart.dart
@@ -1,6 +1,6 @@
 library custom_fl_chart;
 
-export 'src/pie_chart_widget/pie_chart_widget.dart';
-export 'src/data/pie_chart_data_item.dart';
-export 'src/data/balance_entry.dart';
-export 'src/bar_chart_widget/bar_chart_widget.dart';
+export 'src/pie_chart_widget/view/pie_chart_widget.dart';
+export 'src/pie_chart_widget/data/pie_chart_data_item.dart';
+export 'src/bar_chart_widget/data/balance_entry.dart';
+export 'src/bar_chart_widget/view/bar_chart_view.dart';
