+++ packages/custom_fl_chart/lib/src/bar_chart_widget/data/balance_entry.dart
