+++ packages/custom_fl_chart/lib/src/widgets/utils.dart
@@ -1,8 +1,6 @@
 import 'dart:math';
 import 'dart:ui';
 
-
-
 final Map<String, Color> _categoryColorCache = {};
 
 Color getColorForTitle(String title) {
