+++ packages/custom_fl_chart/pubspec.yaml
@@ -13,6 +13,8 @@ dependencies:
 
   fl_chart: ^1.0.0
   intl: any
+  flutter_bloc: any
+  equatable: ^2.0.5
 
 dev_dependencies:
   flutter_test:
